# paas-mediation-service/docs/swagger.yaml
definitions:
...
  v2.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  v2.Metadata:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      generation:
        type: integer
      kind:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      resourceVersion:
        type: string
      uid:
        type: string
    type: object
...
  v2.Service:
    properties:
      metadata:
        $ref: '#/definitions/v2.Metadata'
      spec:
        $ref: '#/definitions/v2.ServiceSpec'
    type: object
  v2.ServicePort:
    properties:
      name:
        type: string
      nodePort:
        type: integer
      port:
        type: integer
      protocol:
        type: string
      targetPort:
        type: integer
    type: object
  v2.ServiceSpec:
    properties:
      clusterIP:
        type: string
      ports:
        items:
          $ref: '#/definitions/v2.ServicePort'
        type: array
      selector:
        additionalProperties:
          type: string
        type: object
      type:
        type: string
    type: object
...
info:
  description: API for Paas Mediation.
  title: Paas Mediation API
  version: "2.0"
paths:
...
  /api/v2/namespaces/{namespace}/services:
    get:
      consumes:
      - application/json
      description: Get Service by name and namespace
      operationId: v2-get-service-list
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: query
        name: annotations
        type: string
      - description: resource name
        in: query
        name: labels
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.Service'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Service by name and namespace
      tags:
      - since:2.0
    post:
      consumes:
      - application/json
      description: Create Service in namespace
      operationId: v2-create-service
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Service'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Service in namespace
      tags:
      - since:2.0
    put:
      consumes:
      - application/json
      description: Update or Create Service in namespace
      operationId: v2-update-or-create-service
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Service'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update or Create Service in namespace
      tags:
      - since:2.0
  /api/v2/namespaces/{namespace}/services/{name}:
    delete:
      consumes:
      - application/json
      description: Delete Service with name in namespace
      operationId: v2-delete-service
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Service'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Service with name in namespace
      tags:
      - since:2.0
    get:
      consumes:
      - application/json
      description: Get Service by name and namespace
      operationId: v2-get-service
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Service'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Service by name and namespace
      tags:
      - since:2.0
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...